# internal/server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Upload Assistant API",
    "version": "2.0.0-dev",
    "description": "Local-first Chinese PT retorrent automation with durable steps, rule gates, and auditable evidence."
  },
  "servers": [{"url": "http://127.0.0.1:8080"}],
  "security": [{"bearerAuth": []}],
  "paths": {
...
    "/api/v2/sites/{site_code}/rule-sources": {
      "get": {
        "operationId": "getSiteRuleSources",
        "description": "Read the ordered HTTPS rule-page source set, per-source authentication mode, and redacted Cookie-presence metadata. Secret values and captured rule text are never returned.",
        "parameters": [{"$ref": "#/components/parameters/SiteCode"}],
        "responses": {"200": {"description": "Configured source set", "content": {"application/json": {"schema": {"type": "object", "properties": {"source_set": {"$ref": "#/components/schemas/RuleSourceSet"}}}}}}, "404": {"$ref": "#/components/responses/Problem"}}
      },
      "put": {
        "operationId": "putSiteRuleSources",
        "description": "Save 1..20 exact HTTPS rule-page URLs, their evidence IDs/scopes, per-source none/site_cookie authentication mode, and explicit completeness/Cookie-host confirmations. Cookie-host confirmation is required only for site_cookie sources. This performs no external call.",
        "parameters": [{"$ref": "#/components/parameters/SiteCode"}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RuleSourceSetInput"}}}},
        "responses": {"200": {"description": "Source set saved and audited", "content": {"application/json": {"schema": {"type": "object", "properties": {"source_set": {"$ref": "#/components/schemas/RuleSourceSet"}}}}}}, "400": {"$ref": "#/components/responses/Problem"}, "409": {"$ref": "#/components/responses/Problem"}}
      }
    },
    "/api/v2/sites/{site_code}/rule-collection-runs": {
      "post": {
        "operationId": "createSiteRuleCollectionRun",
        "description": "Explicitly queue one manual source refresh. Public sources are fetched without credentials; site_cookie sources decrypt and send the existing Cookie only to confirmed exact hosts. The worker shares the normal interval/quota/cooldown/concurrency policy, stores bounded normalized evidence, streams provider inference, imports an immutable draft, and never approves or activates it.",
        "parameters": [{"$ref": "#/components/parameters/SiteCode"}, {"$ref": "#/components/parameters/IdempotencyKey"}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "additionalProperties": false, "required": ["source_set_fingerprint", "provider_id", "confirm"], "properties": {"source_set_fingerprint": {"type": "string", "pattern": "^[a-f0-9]{64}$"}, "provider_id": {"type": "string", "format": "uuid"}, "confirm": {"type": "boolean", "const": true}}}}}},
        "responses": {"202": {"description": "Collection queued", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RuleCollectionEnvelope"}}}}, "400": {"$ref": "#/components/responses/Problem"}, "409": {"$ref": "#/components/responses/Problem"}}
      }
    },
    "/api/v2/sites/{site_code}/rule-collection-runs/latest": {
      "get": {"operationId": "getLatestSiteRuleCollectionRun", "parameters": [{"$ref": "#/components/parameters/SiteCode"}], "responses": {"200": {"description": "Latest run or status=empty", "content": {"application/json": {"schema": {"type": "object"}}}}, "404": {"$ref": "#/components/responses/Problem"}}}
    },
    "/api/v2/site-rule-collection-runs/{run_id}": {
      "get": {"operationId": "getSiteRuleCollectionRun", "parameters": [{"name": "run_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}], "responses": {"200": {"description": "Collection status and bounded source evidence", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RuleCollectionEnvelope"}}}}, "404": {"$ref": "#/components/responses/Problem"}}}
    },
    "/api/v2/site-rule-collection-runs/{run_id}/stream": {
      "get": {"operationId": "streamSiteRuleCollectionRun", "description": "SSE status stream with heartbeats; it observes the durable worker and never starts another collection or provider call.", "parameters": [{"name": "run_id", "in": "path", "required": true, "schema": {"type": "string", "format": "uuid"}}], "responses": {"200": {"description": "SSE progress stream", "content": {"text/event-stream": {"schema": {"type": "string"}}}}, "404": {"$ref": "#/components/responses/Problem"}}}
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {"type": "http", "scheme": "bearer", "bearerFormat": "ua_API_token"}
    },
    "parameters": {
...
      "SiteCode": {"name": "site_code", "in": "path", "required": true, "schema": {"type": "string", "pattern": "^[A-Z0-9_-]+$"}},
...
      "IdempotencyKey": {"name": "Idempotency-Key", "in": "header", "required": true, "schema": {"type": "string", "minLength": 1, "maxLength": 200}}
    },
    "responses": {
      "Problem": {"description": "Structured API problem", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Problem"}}}}
    },
    "schemas": {
...
      "RuleSource": {"type":"object","additionalProperties":false,"required":["id","url","scope"],"properties":{"id":{"type":"string","pattern":"^[a-z0-9][a-z0-9_-]{0,63}$"},"url":{"type":"string","format":"uri","pattern":"^https://","maxLength":2048},"scope":{"type":"string","minLength":1,"maxLength":500},"auth_mode":{"type":"string","enum":["none","site_cookie"],"default":"site_cookie","description":"Authentication for this exact source. none sends no site credential; site_cookie sends only the encrypted site Cookie."}}},
      "RuleSourceSetInput": {"type":"object","additionalProperties":false,"required":["sources","scope_confirmed","cookie_hosts_confirmed"],"properties":{"sources":{"type":"array","minItems":1,"maxItems":20,"items":{"$ref":"#/components/schemas/RuleSource"}},"scope_confirmed":{"type":"boolean"},"cookie_hosts_confirmed":{"type":"boolean"}}},
      "RuleSourceSet": {"allOf":[{"$ref":"#/components/schemas/RuleSourceSetInput"},{"type":"object","required":["site_code","fingerprint","cookie_configured","cookie_required"],"properties":{"site_code":{"type":"string"},"fingerprint":{"type":"string","pattern":"^[a-f0-9]{64}$"},"cookie_configured":{"type":"boolean"},"cookie_required":{"type":"boolean","description":"True only when at least one source uses site_cookie."},"updated_at":{"type":"string","format":"date-time"}}}]},
      "RuleCollectionDocument": {"type":"object","required":["id","source_id","url","scope","auth_mode","status"],"properties":{"id":{"type":"string","format":"uuid"},"source_id":{"type":"string"},"url":{"type":"string","format":"uri"},"scope":{"type":"string"},"auth_mode":{"type":"string","enum":["none","site_cookie"]},"status":{"type":"string","enum":["pending","fetching","ready","failed"]},"http_status":{"type":"integer"},"content_type":{"type":"string"},"size_bytes":{"type":"integer","format":"int64"},"text_sha256":{"type":"string","pattern":"^[a-f0-9]{64}$"},"error_code":{"type":"string"},"error_detail":{"type":"string"},"captured_at":{"type":"string","format":"date-time"}}},
      "RuleCollectionRun": {"type":"object","required":["id","site_code","source_set_fingerprint","provider_id","status","not_before","documents","created_at","updated_at"],"properties":{"id":{"type":"string","format":"uuid"},"site_code":{"type":"string"},"source_set_fingerprint":{"type":"string","pattern":"^[a-f0-9]{64}$"},"provider_id":{"type":"string","format":"uuid"},"provider_config_sha256":{"type":"string","pattern":"^[a-f0-9]{64}$","description":"Non-secret fingerprint of the exact endpoint/model/protocol/data-boundary/credential-version contract bound when queued."},"status":{"type":"string","enum":["queued","fetching","analyzing","ready","failed"]},"not_before":{"type":"string","format":"date-time"},"rule_revision_id":{"type":"string","format":"uuid"},"error_code":{"type":"string"},"error_detail":{"type":"string"},"documents":{"type":"array","items":{"$ref":"#/components/schemas/RuleCollectionDocument"}},"created_at":{"type":"string","format":"date-time"},"started_at":{"type":"string","format":"date-time"},"completed_at":{"type":"string","format":"date-time"},"updated_at":{"type":"string","format":"date-time"}}},
      "RuleCollectionEnvelope": {"type":"object","required":["ok","status","run_id","run","blockers","next_actions"],"properties":{"ok":{"type":"boolean"},"status":{"type":"string"},"run_id":{"type":"string","format":"uuid"},"rule_revision_id":{"type":"string","format":"uuid"},"run":{"$ref":"#/components/schemas/RuleCollectionRun"},"blockers":{"type":"array","items":{}},"next_actions":{"type":"array","items":{}}}},
...
      "Blocker": {
        "type": "object",
        "required": ["code", "message"],
        "properties": {"code": {"type": "string"}, "message": {"type": "string"}, "site_code": {"type": "string"}, "obligation_id": {"type": "string"}},
        "additionalProperties": true
      },
      "NextAction": {
        "type": "object",
        "required": ["action"],
        "properties": {"action": {"type": "string"}, "description": {"type": "string"}, "parameters": {"type": "object", "additionalProperties": true}},
        "additionalProperties": true
      },
...
      "Problem": {"type": "object", "required": ["ok", "status", "error", "blockers", "next_actions"], "properties": {"ok": {"type": "boolean", "const": false}, "status": {"type": "string", "const": "failed"}, "error": {"type": "object", "required": ["code", "detail"], "properties": {"code": {"type": "string"}, "detail": {"type": "string"}}}, "blockers": {"type": "array", "items": {"$ref": "#/components/schemas/Blocker"}}, "next_actions": {"type": "array", "items": {"$ref": "#/components/schemas/NextAction"}}}}
    }
  }
}